Moved debounce function to own util file
